Move files again, use Prototype1 namespace
