+++ CasperVanBattum/Breakin/Assets/Scripts/Prototype1/BasicBlock.cs
@@ -1,0 +1,18 @@
+﻿namespace Prototype1
+{
+	public class BasicBlock : Block
+	{
+		private int strength = 1;
+
+		protected override void OnHit()
+		{
+			strength--;
+
+			// Destroy the gameobject when this block has no more hitpoints
+			if (strength <= 0)
+			{
+				Destroy(gameObject);
+			}
+		}
+	}
+}
