+++ CasperVanBattum/Breakin/Assets/Scripts/Prototype1/Block.cs
@@ -1,0 +1,18 @@
+﻿using UnityEngine;
+
+namespace Prototype1
+{
+	public abstract class Block : MonoBehaviour
+	{
+		/// <summary>
+		/// This function is called each time the ball hits a block
+		/// </summary>
+		protected abstract void OnHit();
+
+		private void OnCollisionEnter2D(Collision2D other)
+		{
+			if (other.gameObject.CompareTag("Ball"))
+				OnHit();
+		}
+	}
+}
